consolidate sending transactions
