+++ packages/app-extension/src/components/Unlocked/Balances/TokensWidget/Solana/index.tsx
@@ -1,33 +1,19 @@
-import { useState } from "react";
-import {
-  Blockchain,
-  confirmTransaction,
-  getLogger,
-  isCardinalWrappedToken,
-  isCreatorStandardToken,
-  isOpenCreatorProtocol,
-  isProgrammableNftToken,
-  SOL_NATIVE_MINT,
-  Solana,
-} from "@coral-xyz/common";
+import { Blockchain } from "@coral-xyz/common";
 import { PrimaryButton, UserIcon } from "@coral-xyz/react-common";
 import {
   useActiveWallet,
   useAvatarUrl,
   useSolanaCtx,
-  useSolanaTokenMint,
+  useSolanaTransaction,
 } from "@coral-xyz/recoil";
 import { styles, useCustomTheme } from "@coral-xyz/themes";
 import { Typography } from "@mui/material";
-import { PublicKey } from "@solana/web3.js";
 import type { BigNumber } from "ethers";
 
 import { CopyablePublicKey } from "../../../../common/CopyablePublicKey";
 import { SettingsList } from "../../../../common/Settings/List";
 import { TokenAmountHeader } from "../../../../common/TokenAmountHeader";
 import { Error, Sending } from "../Send";
-
-const logger = getLogger("send-solana-confirmation-card");
 
 const useStyles = styles((theme) => ({
   confirmTableListItem: {
@@ -63,119 +49,14 @@ export function SendSolanaConfirmationCard({
   onComplete?: (txSig?: any) => void;
   onViewBalances?: () => void;
 }) {
-  const [txSignature, setTxSignature] = useState<string | null>(null);
-  const solanaCtx = useSolanaCtx();
-  const [error, setError] = useState(
-    "Error 422. Transaction time out. Runtime error. Reticulating splines."
-  );
-  const [cardType, setCardType] = useState<
-    "confirm" | "sending" | "complete" | "error"
-  >("confirm");
-  const mintInfo = useSolanaTokenMint({
-    publicKey: solanaCtx.walletPublicKey.toString(),
-    tokenAddress: token.address,
+  const { txSignature, onConfirm, cardType, error } = useSolanaTransaction({
+    token,
+    destinationAddress,
+    amount,
+    onComplete: (txid) => {
+      onComplete?.(txid);
+    },
   });
-
-  const onConfirm = async () => {
-    setCardType("sending");
-    //
-    // Send the tx.
-    //
-    let txSig;
-
-    try {
-      const mintId = new PublicKey(token.mint?.toString() as string);
-      if (token.mint === SOL_NATIVE_MINT.toString()) {
-        txSig = await Solana.transferSol(solanaCtx, {
-          source: solanaCtx.walletPublicKey,
-          destination: new PublicKey(destinationAddress),
-          amount: amount.toNumber(),
-        });
-      } else if (
-        await isProgrammableNftToken(
-          solanaCtx.connection,
-          token.mint?.toString() as string
-        )
-      ) {
-        txSig = await Solana.transferProgrammableNft(solanaCtx, {
-          destination: new PublicKey(destinationAddress),
-          mint: new PublicKey(token.mint!),
-          amount: amount.toNumber(),
-          decimals: token.decimals,
-          source: new PublicKey(token.address),
-        });
-      }
-      // Use an else here to avoid an extra request if we are transferring sol native mints.
-      else {
-        const ocpMintState = await isOpenCreatorProtocol(
-          solanaCtx.connection,
-          mintId,
-          mintInfo
-        );
-        if (ocpMintState !== null) {
-          txSig = await Solana.transferOpenCreatorProtocol(
-            solanaCtx,
-            {
-              destination: new PublicKey(destinationAddress),
-              amount: amount.toNumber(),
-              mint: new PublicKey(token.mint!),
-            },
-            ocpMintState
-          );
-        } else if (isCreatorStandardToken(mintId, mintInfo)) {
-          txSig = await Solana.transferCreatorStandardToken(solanaCtx, {
-            destination: new PublicKey(destinationAddress),
-            mint: new PublicKey(token.mint!),
-            amount: amount.toNumber(),
-            decimals: token.decimals,
-          });
-        } else if (
-          await isCardinalWrappedToken(solanaCtx.connection, mintId, mintInfo)
-        ) {
-          txSig = await Solana.transferCardinalManagedToken(solanaCtx, {
-            destination: new PublicKey(destinationAddress),
-            mint: new PublicKey(token.mint!),
-            amount: amount.toNumber(),
-            decimals: token.decimals,
-          });
-        } else {
-          txSig = await Solana.transferToken(solanaCtx, {
-            destination: new PublicKey(destinationAddress),
-            mint: new PublicKey(token.mint!),
-            amount: amount.toNumber(),
-            decimals: token.decimals,
-          });
-        }
-      }
-    } catch (err: any) {
-      logger.error("solana transaction failed", err);
-      setError(err.toString());
-      setCardType("error");
-      return;
-    }
-
-    setTxSignature(txSig);
-
-    //
-    // Confirm the tx.
-    //
-    try {
-      await confirmTransaction(
-        solanaCtx.connection,
-        txSig,
-        solanaCtx.commitment !== "confirmed" &&
-          solanaCtx.commitment !== "finalized"
-          ? "confirmed"
-          : solanaCtx.commitment
-      );
-      setCardType("complete");
-      if (onComplete) onComplete(txSig);
-    } catch (err: any) {
-      logger.error("unable to confirm", err);
-      setError(err.toString());
-      setCardType("error");
-    }
-  };
 
   return (
     <>
